requirements.txt
flask
flask-cors
scikit-learn
joblib
tldextract
requests
python-Levenshtein